% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{calc_total_return}
\alias{calc_total_return}
\title{Calculate Total Return}
\usage{
calc_total_return(returns)
}
\arguments{
\item{returns}{Numeric vector of returns}
}
\value{
Numeric value representing total return as a decimal (e.g., 0.25 = 25\%)
}
\description{
Calculate the total return of a strategy.
}
\examples{
\dontrun{
returns <- c(0.01, 0.02, -0.01, 0.03)
calc_total_return(returns)
}

}
