% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.performance_summary}
\alias{print.performance_summary}
\title{Print Performance Summary}
\usage{
\method{print}{performance_summary}(x, ...)
}
\arguments{
\item{x}{A performance_summary object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Performance Summary
}
